goodbye firestore

+++ lib/services/user_service.dart
@@ -5,13 +5,15 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_storage/firebase_storage.dart';
+import 'package:point_of_view/constants.dart';
+import 'package:point_of_view/locator.dart';
+import 'package:point_of_view/managers/user_manager.dart';
 import 'package:point_of_view/models/User.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:http/http.dart' as http;
 
 abstract class UserService {
   Future<UserAccount> getUserInfo();
-  Future<UserAccount> getUserInfoList();
   Future<String> uploadProfileImg(String image);
   Future<String> deleteAccount();
   Future<int> getUserIdFromSharedPrefs();
@@ -19,11 +21,10 @@ abstract class UserService {
 }
 
 class UserServiceImplementation implements UserService {
-  var host = 'http://localhost:3000/';
+  // var host = 'http://localhost:3000/';
   // var host = "https://hidden-woodland-36838.herokuapp.com/";
   @override
   Future<UserAccount> getUserInfo() async {
-
     SharedPreferences prefs = await SharedPreferences.getInstance();
     int userId = prefs.getInt('userId');
 
@@ -41,24 +42,23 @@ class UserServiceImplementation implements UserService {
     }
   }
 
-  Future<UserAccount> getUserInfoList() async {
-    String userId = FirebaseAuth.instance.currentUser.uid;
-    DocumentSnapshot doc =
-        await FirebaseFirestore.instance.collection("users").doc(userId).get();
-    return UserAccount.fromJson(doc.data());
-  }
-
   @override
   Future<String> uploadProfileImg(String imagePath) async {
     File file = File(imagePath);
-    String userId = FirebaseAuth.instance.currentUser.uid;
+    int userId = await locator<UserService>().getUserIdFromSharedPrefs();
     String photoUrl = '';
     try {
       TaskSnapshot result = await FirebaseStorage.instance
-          .ref('user_profile_images/$userId/profile_img')
+          .ref('user_profile_images/${userId.toString()}/profile_img')
           .putFile(file);
-      //result.ref.getDownloadURL().then((value) => {photoUrl = value});
       photoUrl = await result.ref.getDownloadURL();
+      var url = host + 'updateProfileImg';
+      var response = await http.put(url,
+          body: {'profileImageUrl': photoUrl, 'userId': userId.toString()});
+      await locator<UserManager>().getUserInfo();
+      if (response.statusCode != 200) {
+        throw Exception('Could not upload profile img');
+      }
     } on FirebaseException catch (e) {
       print(e.code);
     }
@@ -67,11 +67,18 @@ class UserServiceImplementation implements UserService {
 
   Future<String> deleteAccount() async {
     try {
-      await FirebaseFirestore.instance
-          .collection("users")
-          .doc(FirebaseAuth.instance.currentUser.uid)
-          .delete();
-      await FirebaseAuth.instance.currentUser.delete();
+      int userId = await locator<UserService>().getUserIdFromSharedPrefs();
+      var url = host + 'deleteAccount/${userId.toString()}';
+      var response = await http.delete(url);
+      if (response.statusCode != 200)
+        throw Exception('Could not delete user');
+      else {
+        await FirebaseAuth.instance.currentUser.delete();
+        SharedPreferences prefs = await SharedPreferences.getInstance();
+        prefs.setBool('isLoggedIn', false);
+
+        return 'sucess';
+      }
     } catch (e) {
       if (e.code == 'requires-recent-login') {
         print(
